Refatorar handlers de App.jsx para módulos de utilitários
- Move as funções de exportação (handleExportCSV, handleExportHtml) para `src/utils/exportUtils.js`.
- Move a função de download de CSV de exemplo (handleDownloadExampleCSV) para `src/utils/fileUtils.js`.
- Move a função de parsing de resposta da IA (parseIaResponseToCsvData) para `src/utils/iaResponseParser.js`.
- Atualiza App.jsx para importar e usar as novas funções, limpando o componente principal e melhorando a separação de responsabilidades.
- A função handleResetCampaign foi mantida em App.jsx por ser fortemente acoplada ao estado do componente.
