+++ src/App.jsx
@@ -107,6 +107,9 @@ import {
   generateIAContent,
 } from './utils/generationHandlers.js';
 import { saveCampaignState, loadCampaignState } from './utils/campaignState.js';
+import { exportCsv, exportHtml } from './utils/exportUtils.js';
+import { downloadExampleCsv } from './utils/fileUtils.js';
+import { parseIaResponseToCsvData } from './utils/iaResponseParser.js';
 
 // Temas atualizados com gradientes e cores modernas
 const lightTheme = createTheme({
@@ -1023,33 +1026,6 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
     setAnchorElMenu(null);
   };
 
-  const handleDownloadExampleCSV = useCallback(async () => {
-    try {
-      const response = await fetch("/exemplo_posts.csv");
-      if (!response.ok) {
-        throw new Error(`HTTP error! status: ${response.status}`);
-      }
-      const csvText = await response.text();
-
-      // Adicionar BOM UTF-8
-      const csvWithBOM = "\uFEFF" + csvText;
-
-      const blob = new Blob([csvWithBOM], { type: "text/csv;charset=utf-8;" });
-      const url = URL.createObjectURL(blob);
-
-      const link = document.createElement("a");
-      link.href = url;
-      link.download = "exemplo_posts.csv";
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      URL.revokeObjectURL(url);
-    } catch (error) {
-      console.error("Erro ao baixar o CSV de exemplo:", error);
-      alert("Não foi possível baixar o arquivo CSV de exemplo. Verifique o console para mais detalhes.");
-    }
-  }, []);
-
   const handleLoadTemplateClick = () => {
     handleMenuClose();
     // Acionar o clique no input de arquivo escondido
@@ -1061,36 +1037,6 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
   const handleSaveTemplateClick = () => {
     handleMenuClose();
     handleSaveState();
-  };
-
-  const handleExportCSV = () => {
-    if (csvData.length === 0) {
-      alert("Não há dados para exportar.");
-      return;
-    }
-
-    // Papa.unparse espera um array de objetos ou um array de arrays.
-    // Se csvHeaders for usado, ele garante a ordem das colunas.
-    // Se csvData já for um array de objetos com as chaves corretas,
-    // Papa.unparse(csvData) pode ser suficiente, mas usar 'fields' garante a ordem.
-    const config = {
-      quotes: true, // Adiciona aspas em todos os campos
-      delimiter: ";", // Usa ponto e vírgula como delimitador
-      header: true, // Inclui a linha de cabeçalho
-      fields: csvHeaders // Garante a ordem das colunas e quais incluir
-    };
-    const csvString = Papa.unparse(csvData, config);
-
-    const blob = new Blob([`\uFEFF${csvString}`], { type: "text/csv;charset=utf-8;" }); // Adiciona BOM para UTF-8 Excel
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = "dados_exportados.csv";
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-    handleMenuClose(); // Fechar o menu após a ação
   };
   const handleDadosAlterados = useCallback((novosRegistros, novasColunas) => {
     setCsvData(novosRegistros);
@@ -1299,76 +1245,6 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
     setFollowupPostsQuantity(5);
   };
 
-  const handleExportHtml = () => {
-    if (!campaignContent) return;
-
-    const { titulo, conteudo, cta, hashtags } = campaignContent;
-    const imageHtml = backgroundImage ? `
-      <h2>Imagem de Fundo</h2>
-      <img src="${backgroundImage}" alt="Imagem de Fundo da Campanha" style="max-width: 100%; border-radius: 8px; margin-bottom: 2rem;" />
-    ` : '';
-
-    const followupPostsHtml = followupPosts.length > 0 ? `
-      <h2>Posts de Follow-up</h2>
-      ${followupPosts.map(post => `
-        <div style="border: 1px solid #eee; padding: 1rem; margin-bottom: 1rem; border-radius: 8px;">
-          <h3>Post ${post.post_numero}: ${post.tipo_gancho}</h3>
-          <p>${post.conteudo}</p>
-          <p><strong>CTA:</strong> ${post.cta}</p>
-          <div>
-            ${post.hashtags_sugeridas.map(tag => `<span style="background-color: #f5f3ff; color: #6d28d9; padding: 0.25rem 0.75rem; border-radius: 16px; font-size: 0.9rem; margin-right: 0.5rem;">${tag}</span>`).join('')}
-          </div>
-        </div>
-      `).join('')}
-    ` : '';
-
-    const html = `
-      <!DOCTYPE html>
-      <html lang="pt-BR">
-      <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Relatório da Campanha: ${titulo}</title>
-        <style>
-          body { font-family: sans-serif; line-height: 1.6; color: #333; max-width: 800px; margin: 2rem auto; padding: 0 1rem; }
-          h1, h2 { color: #8b5cf6; }
-          .container { border: 1px solid #ddd; border-radius: 8px; padding: 2rem; box-shadow: 0 4px 8px rgba(0,0,0,0.1); }
-          .hashtags { display: flex; flex-wrap: wrap; gap: 0.5rem; margin-top: 1rem; }
-          .hashtag { background-color: #f5f3ff; color: #6d28d9; padding: 0.25rem 0.75rem; border-radius: 16px; font-size: 0.9rem; }
-        </style>
-      </head>
-      <body>
-        <div class="container">
-          <h1>${titulo}</h1>
-          ${imageHtml}
-          <h2>Conteúdo</h2>
-          <p>${conteudo.replace(/\n/g, '<br>')}</p>
-          <h2>Chamada para Ação (CTA)</h2>
-          <p>${cta.replace(/\n/g, '<br>')}</p>
-          <h2>Hashtags</h2>
-          <div class="hashtags">
-            ${hashtags.map(tag => `<span class="hashtag">${tag}</span>`).join('')}
-          </div>
-          ${conteudoMedio ? `<h2>Conteúdo Médio</h2><p>${conteudoMedio.replace(/\n/g, '<br>')}</p>` : ''}
-          ${conteudoPequeno ? `<h2>Conteúdo Pequeno</h2><p>${conteudoPequeno.replace(/\n/g, '<br>')}</p>` : ''}
-          ${conteudoFormatado ? `<h2>Conteúdo Formatado</h2><div>${conteudoFormatado}</div>` : ''}
-          ${followupPostsHtml}
-        </div>
-      </body>
-      </html>
-    `;
-
-    const blob = new Blob([html], { type: 'text/html' });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = `campanha-${titulo.toLowerCase().replace(/\s+/g, '-')}.html`;
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-    URL.revokeObjectURL(url);
-  };
-
   const handleGenerateIAContent = async () => {
     setIsGenerating(true);
     try {
@@ -1409,120 +1285,6 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
       toast.error(`Erro ao gerar conteúdo com IA: ${error.message}`);
     } finally {
       setIsGenerating(false);
-    }
-  };
-
-  const parseIaResponseToCsvData = (responseText) => {
-    // Definição dos cabeçalhos esperados pelo GerenciadorRegistros
-    const finalHeaders = ["Título", "Texto Principal", "Ponte para o Próximo"];
-    const data = [];
-
-    if (!responseText || typeof responseText !== 'string') {
-      console.error("[parseIaResponseToCsvData] Resposta da IA inválida ou vazia.");
-      return { data: [], headers: finalHeaders };
-    }
-
-    console.log("[parseIaResponseToCsvData] Resposta bruta recebida para parsing:", responseText);
-
-    // 1. Extrair o bloco CSV
-    const csvBlockRegex = /```csv\s*([\s\S]+?)\s*```/;
-    const csvMatch = responseText.match(csvBlockRegex);
-    console.log("[parseIaResponseToCsvData] Resultado do match da regex (csvMatch):", csvMatch);
-
-    if (csvMatch && csvMatch[1] && csvMatch[1].trim() !== "") {
-      const csvContent = csvMatch[1].trim();
-      console.log("[parseIaResponseToCsvData] Conteúdo CSV bruto extraído (csvMatch[1]):", csvMatch[1]);
-      console.log("[parseIaResponseToCsvData] Conteúdo CSV após trim (csvContent):", csvContent);
-
-      const parseResult = Papa.parse(csvContent, {
-        header: true,
-        skipEmptyLines: true,
-        dynamicTyping: true,
-      });
-
-      console.log("[parseIaResponseToCsvData] Resultado do Papa.parse:", parseResult);
-
-      if (parseResult.errors && parseResult.errors.length > 0) {
-        console.error("[parseIaResponseToCsvData] Erros durante o parsing com PapaParse:", parseResult.errors.map(err => ({ ...err, input: undefined })));
-      }
-
-      if (parseResult.data && parseResult.data.length > 0) {
-        const actualHeadersFromIA = parseResult.meta.fields || [];
-        console.log("[parseIaResponseToCsvData] Cabeçalhos reais detectados pela IA (via PapaParse):", actualHeadersFromIA);
-
-        const headerMap = {};
-        actualHeadersFromIA.forEach(iaHeader => {
-          const iaHeaderTrimmed = iaHeader.trim();
-          const iaHeaderLower = iaHeaderTrimmed.toLowerCase();
-          if (iaHeaderLower.includes('titulo') || iaHeaderLower.includes('título')) headerMap[iaHeaderTrimmed] = "Título";
-          else if (iaHeaderLower.includes('texto_principal') || iaHeaderLower.includes('texto principal')) headerMap[iaHeaderTrimmed] = "Texto Principal";
-          else if (iaHeaderLower.includes('ponte_proximo') || iaHeaderLower.includes('ponte para o próximo')) headerMap[iaHeaderTrimmed] = "Ponte para o Próximo";
-          else if (iaHeaderLower.includes('id_elemento') || iaHeaderLower.includes('id') || iaHeaderLower.includes('num_slide') || iaHeaderLower.includes('elemento')) headerMap[iaHeaderTrimmed] = "id";
-        });
-        console.log("[parseIaResponseToCsvData] Mapa de Cabeçalhos construído:", headerMap);
-
-        parseResult.data.forEach(rawRecord => {
-          const record = {};
-          let hasTitle = false;
-          for (const iaHeaderMapped in headerMap) {
-            const targetAppHeader = headerMap[iaHeaderMapped];
-            if (Object.prototype.hasOwnProperty.call(rawRecord, iaHeaderMapped)) {
-              let value = rawRecord[iaHeaderMapped];
-              record[targetAppHeader] = value !== null && value !== undefined ? String(value).trim() : "";
-              if (targetAppHeader === "Título" && record[targetAppHeader]) {
-                hasTitle = true;
-              }
-            }
-          }
-          if (hasTitle) {
-            finalHeaders.forEach(appFinalHeader => {
-              if (!record[appFinalHeader]) record[appFinalHeader] = "";
-            });
-            data.push(record);
-          } else {
-            console.warn("[parseIaResponseToCsvData] Registro ignorado por não ter um 'Título' mapeado:", rawRecord);
-          }
-        });
-        console.log("[parseIaResponseToCsvData] Dados Parseados com Sucesso (Gemini CSV via PapaParse):", data);
-        return { data, headers: finalHeaders };
-      } else {
-        console.error("[parseIaResponseToCsvData] PapaParse não retornou dados ou dados eram vazios, mesmo após encontrar bloco CSV.");
-      }
-    } else {
-      console.error("[parseIaResponseToCsvData] Bloco CSV não encontrado ou vazio na resposta da IA. Detalhes do csvMatch:", csvMatch);
-    }
-
-    // Se chegou aqui, o parsing do bloco CSV falhou ou não havia bloco CSV. Tentar fallback.
-    console.log("[parseIaResponseToCsvData] Tentando parser de fallback (formato DeepSeek).");
-    const fallbackLines = responseText.split('\n').map(line => line.trim()).filter(line => line.length > 0);
-    let currentRecord = {};
-    const fallbackData = []; // Usar um novo array para o fallback
-
-    for (const line of fallbackLines) {
-      if (line.toLowerCase().startsWith("título:") || line.toLowerCase().startsWith("titulo:")) {
-        if (Object.keys(currentRecord).length > 0 && currentRecord["Título"]) fallbackData.push(currentRecord);
-        currentRecord = { "Título": line.substring(line.indexOf(':') + 1).trim() };
-      } else if (line.toLowerCase().startsWith("texto principal:")) {
-        currentRecord["Texto Principal"] = line.substring(line.indexOf(':') + 1).trim();
-      } else if (line.toLowerCase().startsWith("ponte para o próximo:") || line.toLowerCase().startsWith("ponte:")) {
-        currentRecord["Ponte para o Próximo"] = line.substring(line.indexOf(':') + 1).trim();
-        if (currentRecord["Título"]) fallbackData.push(currentRecord);
-        currentRecord = {};
-      }
-    }
-    if (Object.keys(currentRecord).length > 0 && currentRecord["Título"]) fallbackData.push(currentRecord);
-
-    if (fallbackData.length > 0) {
-      console.log("[parseIaResponseToCsvData] Parseado como fallback (formato DeepSeek):", JSON.parse(JSON.stringify(fallbackData)));
-      const processedData = fallbackData.map(record => ({
-        "Título": record["Título"] || "",
-        "Texto Principal": record["Texto Principal"] || "",
-        "Ponte para o Próximo": record["Ponte para o Próximo"] || "",
-      }));
-      return { data: processedData, headers: finalHeaders };
-    } else {
-      console.error("[parseIaResponseToCsvData] Fallback também não encontrou dados estruturados.");
-      return { data: [], headers: finalHeaders }; // Retorna data vazia se tudo falhar
     }
   };
 
@@ -1650,7 +1412,7 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
                   <FileUploadIcon sx={{ mr: 1 }} />
                   Carregar Campanha
                 </MenuItem>
-                <MenuItem onClick={handleExportCSV} disabled={csvData.length === 0}>
+                <MenuItem onClick={() => { exportCsv(csvData, csvHeaders); handleMenuClose(); }} disabled={csvData.length === 0}>
                   <DownloadIcon sx={{ mr: 1 }} />
                   Exportar CSV
                 </MenuItem>
@@ -1788,7 +1550,14 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
               campaignContent={campaignContent}
               handleGenerateCampaignContent={handleGenerateCampaignContent}
               handleResetCampaign={handleResetCampaign}
-              handleExportHtml={handleExportHtml}
+              handleExportHtml={() => exportHtml({
+                campaignContent,
+                backgroundImage,
+                followupPosts,
+                conteudoMedio,
+                conteudoPequeno,
+                conteudoFormatado
+              })}
               editingField={editingField}
               setEditingField={setEditingField}
               conteudoMedio={conteudoMedio}
@@ -1895,7 +1664,7 @@ A resposta DEVE ser um único objeto JSON, sem nenhum texto ou formatação mark
                           </Button>
                           <Button
                             variant="contained"
-                            onClick={handleDownloadExampleCSV}
+                            onClick={downloadExampleCsv}
                             sx={{ borderRadius: 2 }}
                           >
                             Baixar CSV Exemplo
